Get app_server working on x86_64.
With this commit, app_server now compiles and runs at boot! Nothing
particularly interesting happens, just the blue background and a mouse
pointer. Remote backends are broken and not compiled in, see #8834.
Note that it won't be possible to build this quite yet, need to get
the FreeType package uploaded.

+++ src/add-ons/input_server/devices/wacom/MasterServerDevice.cpp
@@ -60,7 +60,7 @@ MasterServerDevice::~MasterServerDevice()
 	// cleanup
 	_StopAll();
 	if (_LockDevices()) {
-		while (PointingDevice* device = (PointingDevice*)fDevices.RemoveItem(0L))
+		while (PointingDevice* device = (PointingDevice*)fDevices.RemoveItem((int32)0))
 			delete device;
 		_UnlockDevices();
 	}
